Add required queries for patient Carbs & Glucose

+++ back-end/Sacchon/src/main/java/gr/codehub/sacchon/app/dto/CarbsDto.java
@@ -11,12 +11,12 @@ import java.time.LocalDate;
 @AllArgsConstructor
 @Setter
 @Getter
-public class CarbsDto implements Measurement<Double> {
+public class CarbsDto implements Measurement<Integer> {
 
     private int id;
     private final String UNITS = "g";
     private LocalDate date;
-    private Double measurement;
+    private Integer measurement;
     private PatientDto patientDto;
 
     public CarbsDto(Carbs carbs) {
