+++ back-end/Sacchon/src/main/java/gr/codehub/sacchon/app/dto/CarbsFromPersonDto.java
@@ -11,12 +11,12 @@ import java.time.LocalDate;
 @AllArgsConstructor
 @Setter
 @Getter
-public class CarbsFromPersonDto implements Measurement<Double> {
+public class CarbsFromPersonDto implements Measurement<Integer> {
 
     private int id;
     private final String UNITS = "g";
     private LocalDate date;
-    private Double measurement;
+    private Integer measurement;
 
     public CarbsFromPersonDto(Carbs carbs) {
         if (carbs == null)
