+++ back-end/Sacchon/src/main/java/gr/codehub/sacchon/app/model/Carbs.java
@@ -12,14 +12,14 @@ import java.time.LocalDate;
 @Setter
 @Getter
 @Table(schema = SacchonApplication.SCHEMA)
-public class Carbs implements Measurement<Double> {
+public class Carbs implements Measurement<Integer> {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private int id;
     private final String UNITS = "g";
     private LocalDate date;
-    private Double measurement;
+    private Integer measurement;
     @ManyToOne(optional = false)
     @JoinColumn(name = "patient_id", referencedColumnName = "id")
     private Patient patient;
